+++ NEWS.md
@@ -3,6 +3,9 @@
 ## 1.44.2
 Released 2017-mm-dd
 
+Bug fixes:
+ - Update queue index while enqueueing jobs to the top of queue.
+
 
 ## 1.44.1
 Released 2017-04-04